Add new SaveResults reducer for ActiveRuns state
